+++ utils/carousel.ts
@@ -5,6 +5,7 @@ import {
 } from '~/utils/cachingStrategy'
 
 import { CarouselNFT } from '@/components/base/types'
+import { processSingleMetadata } from '~/utils/cachingStrategy'
 
 /**
  * Format the data to fit with CarouselNFT[]
@@ -26,4 +27,16 @@ export const formatNFT = async (data): Promise<CarouselNFT[]> => {
     }),
     image: imageOf(nft.meta.id, nft.meta.image),
   }))
+}
+
+export const fallbackMetaByNftEvent = async (events: any[]) => {
+  for (const event of events) {
+    if (!event.nft.meta) {
+      event.nft.meta = {
+        id: event.nft.metadata,
+        image: '',
+      }
+      await processSingleMetadata(event.nft.metadata)
+    }
+  }
 }
